pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "finance-llm"
version = "0.1.0"
description = "Personal finance system with LLM-assisted insights via hledger"
requires-python = ">=3.11"
dependencies = [
    "click>=8.0",
    "pyyaml>=6.0",
    "rich>=13.0",
    "mcp>=1.0",
    "openai>=1.0",
    "google-api-python-client>=2.0",
    "google-auth-oauthlib>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "ruff>=0.4",
]

[project.scripts]
fin = "finance_llm.bin.fin:main"
fin-ingest = "finance_llm.bin.ingest:main"
fin-review = "finance_llm.bin.review:main"
fin-post = "finance_llm.bin.post:main"
fin-mcp = "finance_llm.bin.mcp:main"

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.pytest.ini_options]
testpaths = ["tests"]
